Move logging methods into malsim_vectorized_obs_env.py
